requirements.txt
# Flask==2.1.3
# psycopg2-binary==2.9.1
Flask==2.1.3
opencv-python-headless==4.5.3.56
psycopg2-binary==2.9.1
numpy==1.21.1
qrcode[pil]==7.3
pyzbar==0.1.8
